Final Game

+++ Services/CharacterService/CharacterService.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using dotnet_webapi_rpg.Data;
@@ -13,32 +14,47 @@ namespace dotnet_webapi_rpg.Services.CharacterService
     {
         private readonly IMapper _mapper;
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public CharacterService(IMapper mapper, DataContext context) {
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
+        {
             _mapper = mapper;
             _context = context;
+            _httpContextAccessor = httpContextAccessor;
         }
-        public async Task<ServiceResponse<List<GetCharacterDto>>> CreateCharacter(AddCharacterDto newCharacter)
+
+        // Get the current logged in user id
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
+
+        public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
         {
             var response = new ServiceResponse<List<GetCharacterDto>>();
             // Map AddCharacterDto to Character
             Character character = _mapper.Map<Character>(newCharacter);
+            // Assign the User field in charachter to the user who created it
+            character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
             //Fetch caracters from the database and add the new character to the list
             _context.Characters.Add(character);
             //Save the changes to the database
             await _context.SaveChangesAsync();
-            // Map every character of type Character in the list to a GetCharacterDto
-            response.Data = await _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+            // Map every character of type Character in the list to a GetCharacterDto for a specific user
+            response.Data = await _context.Characters
+                .Where(c => c.User.Id == GetUserId())
+                .Select(c => _mapper.Map<GetCharacterDto>(c))
+                .ToListAsync();
             return response;
         }
 
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             var response = new ServiceResponse<GetCharacterDto>();
-            
+
             try
             {
-                var dbCharacter = await _context.Characters.FirstAsync<Character>(c => c.Id == id);
+                var dbCharacter = await _context.Characters
+                    .Include(c => c.Weapon)
+                    .Include(c => c.Skills)
+                    .FirstAsync<Character>(c => c.Id == id && c.User.Id == GetUserId());
                 response.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
             }
             catch (System.Exception ex)
@@ -50,11 +66,13 @@ namespace dotnet_webapi_rpg.Services.CharacterService
             return response;
         }
 
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetCharacters(int userId)
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetCharacters()
         {
             var response = new ServiceResponse<List<GetCharacterDto>>();
             var dbCharacters = await _context.Characters
-                .Where(c => c.User.Id == userId) // characters for a specific user
+                .Include(c => c.Weapon)
+                .Include(c => c.Skills)
+                .Where(c => c.User.Id == GetUserId()) // characters for a specific user
                 .ToListAsync();
             response.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
             return response;
@@ -63,20 +81,32 @@ namespace dotnet_webapi_rpg.Services.CharacterService
         public async Task<ServiceResponse<GetCharacterDto>> UpdateCharacter(UpdatedCharacterDto updatedCharacter)
         {
             ServiceResponse<GetCharacterDto> response = new ServiceResponse<GetCharacterDto>();
-           
+
             try
             {
-                Character character = await _context.Characters.FirstAsync(c => c.Id == updatedCharacter.Id);
+                Character character = await _context.Characters
+                    // By including the Userlinked to a character, this field can be used outside of the DataContext
+                    .Include(c => c.User)
+                    .FirstAsync(c => c.Id == updatedCharacter.Id);
+                // Check if the specific user can update this character
+                if (character.User.Id == GetUserId())
+                {
+                    // Override ALL the character properties. In case you want to change only a few propoerties, you can
+                    // do something like `characted.Name = updatedCharacter.Name`, and so on
+                    _mapper.Map(updatedCharacter, character);
 
-                // Override ALL the character properties. In case you want to change only a few propoerties, you can
-                // do something like `characted.Name = updatedCharacter.Namen`, and so on
-                _mapper.Map(updatedCharacter, character);
+                    // save the updated character to the database
+                    await _context.SaveChangesAsync();
 
-                // save the updated character to the database
-                await _context.SaveChangesAsync();
+                    // Map the Character to a GetCharacter Dto
+                    response.Data = _mapper.Map<GetCharacterDto>(character);
+                }
+                else
+                {
+                    response.Success = false;
+                    response.Message = "Character not found";
+                }
 
-                // Map the Character to a GetCharacter Dto
-                response.Data = _mapper.Map<GetCharacterDto>(character);
             }
             catch (System.Exception ex)
             {
@@ -93,10 +123,21 @@ namespace dotnet_webapi_rpg.Services.CharacterService
 
             try
             {
-                Character character = await _context.Characters.FirstAsync(c => c.Id == id);
-                _context.Characters.Remove(character);
-                await _context.SaveChangesAsync();
-                response.Data = await _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+                Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+                if (character != null)
+                {
+                    _context.Characters.Remove(character);
+                    await _context.SaveChangesAsync();
+                    response.Data = await _context.Characters
+                        .Where(c => c.User.Id == GetUserId())
+                        .Select(c => _mapper.Map<GetCharacterDto>(c))
+                        .ToListAsync();
+                }
+                else
+                {
+                    response.Success = false;
+                    response.Message = "Character not found";
+                }
             }
             catch (System.Exception ex)
             {
@@ -104,6 +145,47 @@ namespace dotnet_webapi_rpg.Services.CharacterService
                 response.Message = ex.Message;
             }
 
+            return response;
+        }
+
+        public async Task<ServiceResponse<GetCharacterDto>> AddCharacterSkill(AddCharacterSkillDto newCharacterSkill)
+        {
+            ServiceResponse<GetCharacterDto> response = new ServiceResponse<GetCharacterDto>();
+
+            try
+            {
+                Character character = await _context.Characters
+                    .Include(c => c.Skills)
+                    .Include(c => c.Weapon)
+                    .FirstOrDefaultAsync(c => c.Id == newCharacterSkill.CharacterId && c.User.Id == GetUserId()
+                );
+
+                if (character == null)
+                {
+                    response.Success = false;
+                    response.Message = "Character not found";
+                    return response;
+                }
+
+                Skill skill = await _context.Skills
+                    .FirstOrDefaultAsync(s => s.Id == newCharacterSkill.SkillId);
+
+                if (skill == null)
+                {
+                    response.Success = false;
+                    response.Message = "Skill not found";
+                    return response;
+                }
+
+                character.Skills.Add(skill);
+                await _context.SaveChangesAsync();
+                response.Data = _mapper.Map<GetCharacterDto>(character);
+            }
+            catch (System.Exception ex)
+            {
+                response.Success = false;
+                response.Message = ex.Message;
+            }
             return response;
         }
     }
